add UI Kitten Header to StagesMapView
