+++ app/(tabs)/(touren)/_layout.tsx
@@ -35,7 +35,7 @@ export default function ReiseStackLayout() {
       <Stack.Screen
         name="stagesMapViewWrapper"
         options={{
-          headerShown: true,
+          headerShown: false,
           title: "Tourübersicht",
         }}
       />
