+++ app/(tabs)/(touren)/stagesMapViewWrapper.tsx
@@ -1,24 +1,35 @@
 // Wrapper component that extracts `tourId` and passes it as a prop
 import { useLocalSearchParams } from "expo-router";
-import { Layout, Text } from "@ui-kitten/components";
+import { Divider, Layout, Text, TopNavigation } from "@ui-kitten/components";
 import StagesMapView from "../../../components/Tour/StagesMapView";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+import renderBackAction from "@/components/TopNavigation/renderBackAction";
 
 type localParams = {
   tourId?: string;
 };
 
 export default function StagesMapViewWrapper() {
+  const insets = useSafeAreaInsets();
   const { tourId } = useLocalSearchParams<localParams>(); // Get tourId from URL
-  if (!tourId) {
-    return (
-      <Layout>
+
+  // Pass tourId to the enhanced component
+  return (
+    <Layout style={{ flex: 1 }}>
+      <TopNavigation
+        title={() => <Text category="h4">Tourübersicht</Text>}
+        style={{ marginTop: insets.top }}
+        alignment="center"
+        accessoryLeft={renderBackAction}
+      ></TopNavigation>
+      <Divider />
+      {tourId ? (
+        <StagesMapView tourId={tourId} />
+      ) : (
         <Text category="h5" status="danger">
           No tourId provided.
         </Text>
-      </Layout>
-    );
-  }
-
-  // Pass tourId to the enhanced component
-  return <StagesMapView tourId={tourId} />;
+      )}
+    </Layout>
+  );
 }
